Mike & Bret's comment
